+++ packages/biconomy-middleware/src/registry.ts
@@ -1,4 +1,6 @@
+import { handleFetchErrorResponse, NetworkError } from "@rarible/logger/build"
 import type { ContractMetadata, IContractRegistry } from "./types"
+import { NetworkErrorCode } from "./domain"
 
 export class Registry implements IContractRegistry {
 	private registryData!: Record<string, ContractMetadata>
@@ -7,8 +9,18 @@ export class Registry implements IContractRegistry {
 	}
 
 	private async fetchData() {
-		const resp = await fetch(this.registryUrl)
-		this.registryData = await resp.json()
+		let response
+		try {
+		  response = await fetch(this.registryUrl)
+		} catch (e) {
+			throw new NetworkError({
+				url: this.registryUrl,
+				data: (e as Error).message,
+				code: NetworkErrorCode.BICONOMY_EXTERNAL_ERR,
+			})
+		}
+		await handleFetchErrorResponse(response, { code: NetworkErrorCode.BICONOMY_EXTERNAL_ERR })
+		this.registryData = await response.json()
 	}
 
 	async getMetadata(address: string, data?: string): Promise<ContractMetadata | undefined> {