+++ packages/ethers-ethereum/src/utils/parse-logs.ts
@@ -2,11 +2,10 @@ import type { ethers, Contract } from "ethers"
 import type { EthereumTransactionEvent } from "@rarible/ethereum-provider"
 
 export async function getTxEvents(
-	receiptPromise: Promise<ethers.providers.TransactionReceipt>,
+	receipt: ethers.providers.TransactionReceipt,
 	contract: Contract
 ): Promise<EthereumTransactionEvent[]> {
 	try {
-		const receipt = await receiptPromise
 		return receipt.logs.map(log => {
 			try {
 				const parsedEvent = contract.interface.parseLog(log)