+++ packages/sdk/src/common/get-base-fee.ts
@@ -1,9 +1,11 @@
 import type { AxiosResponse } from "axios"
 import axios from "axios"
+import { handleAxiosErrorResponse } from "@rarible/logger/build"
 import type { EthereumConfig } from "../config/type"
 import type { EthereumNetwork } from "../types"
 import type { SimpleOrder } from "../order/types"
 import { CURRENT_ORDER_TYPE_VERSION } from "./order"
+import { NetworkErrorCode } from "./logger/logger"
 
 export async function getBaseFee(
 	config: EthereumConfig,
@@ -16,11 +18,12 @@ export async function getBaseFee(
 		envFeeConfig = commonFeeConfigResponse.data[env]
 
 		if (!envFeeConfig) {
-			throw new Error(`Fee config not found for ${env}`)
+			throw new Error(`Fee config was not found for ${env}`)
 		}
 	} catch (e) {
 		console.error(e)
-		throw new Error("Config with fee variables cannot be loaded")
+		handleAxiosErrorResponse(e, { code: NetworkErrorCode.ETHEREUM_EXTERNAL_ERR })
+		throw e
 	}
 
 	if (!(type in envFeeConfig)) {