+++ packages/sdk/src/config/testnet.ts
@@ -10,7 +10,7 @@ export const testnetEthereumConfig: EthereumConfig = {
 		v1: toAddress("0xda381535565b97640a6453fa7a1a7b161af78cbe"),
 		v2: toAddress("0xd4a57a3bD3657D0d46B4C5bAC12b3F156B9B886b"),
 		openseaV1: toAddress("0xdd54d660178b28f6033a953b0e55073cfa7e3744"),
-		wrapper: toAddress("0xa82471938f61bFC5fF361874ca5A4e4f1C2524d7"),
+		wrapper: toAddress("0x3D830cB13043912af930587a471D6C9b32Aa42E9"),
 		looksrare: toAddress("0x1AA777972073Ff66DCFDeD85749bDD555C0665dA"),
 		x2y2: ZERO_ADDRESS,
 	},
@@ -35,6 +35,10 @@ export const testnetEthereumConfig: EthereumConfig = {
 	cryptoPunks: {
 		marketContract: toAddress("0x85252f525456d3fce3654e56f6eaf034075e231c"),
 		wrapperContract: toAddress("0x7898972f9708358acb7ea7d000ebdf28fcdf325c"),
+	},
+	sudoswap: {
+		pairFactory: toAddress("0xcB1514FE29db064fa595628E0BFFD10cdf998F33"),
+		pairRouter: toAddress("0x9ABDe410D7BA62fA11EF37984c0Faf2782FE39B5"),
 	},
 	weth: toAddress("0xc778417e063141139fce010982780140aa0cd5ab"),
 	auction: ZERO_ADDRESS,