+++ packages/sdk/src/order/fill-order/common/get-updated-call.ts
@@ -5,6 +5,8 @@ import type { IRaribleEthereumSdkConfig } from "../../../types"
 
 
 export function getUpdatedCalldata(sdkConfig?: IRaribleEthereumSdkConfig): Binary | undefined {
+	const callDataLength = 48
+
 	if (sdkConfig?.fillCalldata) {
 		const hexRegexp = /^[0-9a-f]*$/i
 		const fillCalldata = toBinary(sdkConfig.fillCalldata).slice(2).toString()
@@ -12,10 +14,12 @@ export function getUpdatedCalldata(sdkConfig?: IRaribleEthereumSdkConfig): Binar
 		if (isNotHexValue) {
 			throw new Error("Fill calldata is not a hex value")
 		}
-		if (fillCalldata.length !== 48) {
-			throw new Error(`Fill call data has length = ${fillCalldata.length}, but should be = 48`)
+		if (fillCalldata.length !== callDataLength) {
+			throw new Error(`Fill call data has length = ${fillCalldata.length}, but should be = ${callDataLength}`)
 		}
 		return toBinary(`0x${fillCalldata}${FILL_CALLDATA_TAG}`)
+	} else {
+		//use default
+		return toBinary(`0x${"0".repeat(callDataLength)}${FILL_CALLDATA_TAG}`)
 	}
-	return undefined
 }